feat(web): add shared hooks for list pages and adopt in sauces/products/recipes clients
- Add hooks: useDebouncedValue, useUrlStateSync, useFirstPaint
- Replace duplicated logic in list clients (URL sync, debounce, hydration guard)
- No behavior changes; prepares for further dedup (SortDropdown, shared filter primitives)
