+++ apps/web/src/hooks/use-debounced-value.ts
@@ -1,0 +1,13 @@
+"use client";
+import { useEffect, useState } from "react";
+
+export function useDebouncedValue<T>(value: T, delay = 200): T {
+  const [debounced, setDebounced] = useState<T>(value);
+
+  useEffect(() => {
+    const id = setTimeout(() => setDebounced(value), delay);
+    return () => clearTimeout(id);
+  }, [value, delay]);
+
+  return debounced;
+}
